+++ apps/ledger-live-mobile/src/components/RootNavigator/types/BaseNavigator.ts
@@ -59,7 +59,7 @@ import type { CosmosClaimRewardsFlowParamList } from "../../../families/cosmos/C
 import type { SolanaDelegationFlowParamList } from "../../../families/solana/DelegationFlow/types";
 import type { StellarAddAssetFlowParamList } from "../../../families/stellar/AddAssetFlow/types";
 import type { TezosDelegationFlowParamList } from "../../../families/tezos/DelegationFlow/types";
-import type { EditTransactionParamList } from "../../../families/ethereum/EditTransactionParamList";
+import type { EditTransactionParamList } from "../../../families/ethereum/EditTransaction/EditTransactionParamList";
 import type { TronVoteFlowParamList } from "../../../families/tron/VoteFlow/types";
 import type { NoFundsNavigatorParamList } from "./NoFundsNavigator";
 import type { StakeNavigatorParamList } from "./StakeNavigator";
