+++ apps/ledger-live-mobile/src/families/bitcoin/SendRowsFee.tsx
@@ -15,7 +15,7 @@ import type { SendFundsNavigatorStackParamList } from "../../components/RootNavi
 import type { BaseNavigatorStackParamList } from "../../components/RootNavigator/types/BaseNavigator";
 import type { SignTransactionNavigatorParamList } from "../../components/RootNavigator/types/SignTransactionNavigator";
 import type { SwapNavigatorParamList } from "../../components/RootNavigator/types/SwapNavigator";
-import { EditTransactionParamList } from "../ethereum/EditTransactionParamList";
+import { EditTransactionParamList } from "../ethereum/EditTransaction/EditTransactionParamList";
 
 type Props = {
   transaction: Transaction;
