+++ apps/ledger-live-mobile/src/families/ethereum/EditTransaction/EditTransactionNavigator.tsx
@@ -7,7 +7,7 @@ import { ScreenName } from "../../../const";
 import { getStackNavigatorConfig } from "../../../navigation/navigatorConfig";
 import StepHeader from "../../../components/StepHeader";
 import { MethodSelection } from "./MethodSelection";
-import { EditTransactionParamList } from "../EditTransactionParamList";
+import { EditTransactionParamList } from "./EditTransactionParamList";
 import SelectDevice from "../../../screens/SelectDevice";
 import SendConnectDevice from "../../../screens/ConnectDevice";
 import SendSummary from "../../../screens/SendFunds/04-Summary";
