+++ apps/ledger-live-mobile/src/families/ethereum/EditTransaction/EditTransactionParamList.ts
@@ -6,7 +6,7 @@ import { Transaction, TransactionStatus } from "@ledgerhq/live-common/generated/
 import { Device } from "@ledgerhq/types-devices";
 import { Account, AccountLike, Operation } from "@ledgerhq/types-live";
 
-import { ScreenName } from "../../const";
+import { ScreenName } from "../../../const";
 
 export type EditTransactionParamList = {
   [ScreenName.TransactionAlreadyValidatedError]: {
