+++ apps/ledger-live-mobile/src/families/ethereum/types.ts
@@ -8,7 +8,7 @@ import { BaseNavigatorStackParamList } from "../../components/RootNavigator/type
 import { SwapNavigatorParamList } from "../../components/RootNavigator/types/SwapNavigator";
 import { StackNavigatorProps } from "../../components/RootNavigator/types/helpers";
 import { ScreenName } from "../../const";
-import { EditTransactionParamList } from "./EditTransactionParamList";
+import { EditTransactionParamList } from "./EditTransaction/EditTransactionParamList";
 
 export type SendRowsFeeProps<T extends Transaction = Transaction> = {
   transaction: T;
