+++ apps/ledger-live-mobile/src/families/ripple/SendRowsFee.tsx
@@ -9,7 +9,7 @@ import type { SendFundsNavigatorStackParamList } from "../../components/RootNavi
 import { ScreenName } from "../../const";
 import { SignTransactionNavigatorParamList } from "../../components/RootNavigator/types/SignTransactionNavigator";
 import { SwapNavigatorParamList } from "../../components/RootNavigator/types/SwapNavigator";
-import { EditTransactionParamList } from "../ethereum/EditTransactionParamList";
+import { EditTransactionParamList } from "../ethereum/EditTransaction/EditTransactionParamList";
 
 type Props = {
   transaction: Transaction;
