+++ apps/ledger-live-mobile/src/families/stellar/SendRowsCustom.tsx
@@ -7,7 +7,7 @@ import { SendFundsNavigatorStackParamList } from "../../components/RootNavigator
 import { SignTransactionNavigatorParamList } from "../../components/RootNavigator/types/SignTransactionNavigator";
 import { SwapNavigatorParamList } from "../../components/RootNavigator/types/SwapNavigator";
 import { ScreenName } from "../../const";
-import { EditTransactionParamList } from "../ethereum/EditTransactionParamList";
+import { EditTransactionParamList } from "../ethereum/EditTransaction/EditTransactionParamList";
 
 type Navigation = BaseComposite<
   | StackNavigatorProps<SendFundsNavigatorStackParamList, ScreenName.SendSummary>
