+++ apps/ledger-live-mobile/src/screens/SendFunds/04-Summary.tsx
@@ -42,7 +42,7 @@ import { BaseComposite, StackNavigatorProps } from "../../components/RootNavigat
 import { SignTransactionNavigatorParamList } from "../../components/RootNavigator/types/SignTransactionNavigator";
 import { SwapNavigatorParamList } from "../../components/RootNavigator/types/SwapNavigator";
 import { CurrentNetworkFee } from "../../families/ethereum/CurrentNetworkFee";
-import { EditTransactionParamList } from "../../families/ethereum/EditTransactionParamList";
+import { EditTransactionParamList } from "../../families/ethereum/EditTransaction/EditTransactionParamList";
 
 type Navigation = BaseComposite<
   | StackNavigatorProps<SendFundsNavigatorStackParamList, ScreenName.SendSummary>
